Fix broken item indicator logic and update rotation type
Refactored BrokenItemIndicatorRenderer to properly track and update broken models using a dedicated Set. Changed rawRotation type from u8 to u16 in EmpireProfileTemplate to support a wider range of rotation values.

+++ src/shared/data/profile/EmpireProfileTemplate.ts
@@ -45,7 +45,7 @@ declare global {
         /**
          * The rotation value in degrees.
          */
-        rawRotation?: DataType.u8;
+        rawRotation?: DataType.u16;
 
         /**
          * The area of the placement, used to identify the item in the world.
